requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy[asyncio]>=2.0.25
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.4.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.18
email-validator>=2.1.0.post1
Jinja2>=3.1.5
python-dotenv==1.0.0
aiosmtplib>=3.0.1
alembic>=1.13.1
aiosqlite>=0.19.0
openai==1.12.0
httpx==0.24.1
textract>=1.6.5
loguru>=0.7.0
websockets>=11.0.3
PyYAML>=6.0.1
sse-starlette==0.10.3
